Only run when instructed to

+++ bin/travis.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
+if [[ ${RUN_E2E} != 1 ]]; then
+        exit 0
+fi
 
 if [ $1 == 'before' ]; then
 		phpenv config-rm xdebug.ini
